% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LRL.R
\name{LRL}
\alias{LRL}
\title{Low-Rank Linear regression for matrix responses.}
\usage{
LRL(x, y, lambda = 1, epsilon = 0.1)
}
\arguments{
\item{x}{The coefficient matrix in regression.}

\item{y}{The list consisting of  samples' responses.}

\item{lambda}{Penalty factor.}

\item{epsilon}{Iteration limit.}
}
\value{
a list consisting of parameter estimation.
}
\description{
Low-Rank Linear regression for matrix responses.
}
\examples{
\dontrun{
data(xy)
x<-xy$X
y<-xy$Y
B_hat<-LRL(x,y)
}
}
