% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RwmsC}
\alias{RwmsC}
\title{A random walk Metropolis sampler via Rcpp}
\usage{
RwmsC(sigma, x_0, N)
}
\arguments{
\item{sigma}{Standard deviation of normal distribution}

\item{x_0}{The initial value of iteration}

\item{n}{The number of samples}
}
\value{
a random sample of size \code{n}
}
\description{
A random walk Metropolis sampler Rcpp
}
\examples{
\dontrun{
x<-RwmsC(2,5,1e5)
hist(x)
}
}
